+++ xclim/indices/_conversion.py
@@ -54,7 +54,7 @@ def uas_vas_2_sfcwind(
       Eastward wind velocity (m s-1)
     vas : xr.DataArray
       Northward wind velocity (m s-1)
-    return_direction:: bool
+    return_direction: bool
       If true, returns direction, else returns surface windspeed.
 
     Returns