+++ xclim/indices/generic.py
@@ -8,7 +8,7 @@ Helper functions for common generic actions done in the computation of indices.
 """
 # Note: scipy.stats.dist.shapes: comma separated names of shape parameters
 # The other parameters, common to all distribution, are loc and scale.
-from typing import Sequence, Union
+from typing import Optional, Sequence, Union
 
 import dask.array
 import numpy as np
@@ -59,7 +59,7 @@ def select_time(da: xr.DataArray, **indexer):
     return selected
 
 
-def select_resample_op(da: xr.DataArray, op, freq: str = "YS", **indexer):
+def select_resample_op(da: xr.DataArray, op: str, freq: str = "YS", **indexer):
     """Apply operation over each period that is part of the index selection.
 
     Parameters
@@ -305,7 +305,15 @@ def fa(
     return out
 
 
-def frequency_analysis(da, mode, t, dist, window=1, freq=None, **indexer):
+def frequency_analysis(
+    da: xr.DataArray,
+    mode: str,
+    t: Union[int, Sequence[int]],
+    dist: str,
+    window: int = 1,
+    freq: Optional[str] = None,
+    **indexer,
+):
     """Return the value corresponding to a return period.
 
     Parameters
@@ -324,7 +332,7 @@ def frequency_analysis(da, mode, t, dist, window=1, freq=None, **indexer):
       Averaging window length (days).
     freq : str
       Resampling frequency. If None, the frequency is assumed to be 'YS' unless the indexer is season='DJF',
-      in which case `freq` would be set to `YS-DEC`.
+      in which case `freq` would be set to `AS-DEC`.
     **indexer : {dim: indexer, }, optional
       Time attribute and values over which to subset the array. For example, use season='DJF' to select winter values,
       month=1 to select January, or month=[6,7,8] to select summer months. If not indexer is given, all values are